moved to cmake - it's a bit shit, but I'll try to fix it with code generation + started to write a parser
