+++ lib/node.cc
@@ -1,0 +1,1 @@
+// empty
